Major Refactoring of property array methods
[C++] Replaces type-specific property array methods with generic pushBackArray* and getArrayOfValuesOfPatch template methods in example.cpp.
Can now get/set and compute statistics when reading/writing these arrays (RESQML2.0.1 only support min and max statistics).
[SWIG] Rename pushBack*Hdf5ArrayOfValues by pushBack*ArrayOfValues (i.e remove Hdf5)

+++ src/resqml2/ContinuousProperty.h
@@ -230,7 +230,7 @@ namespace RESQML2_NS
 		 * 			greater than @p 0 for a non vector property or greater than the vector size for a
 		 * 			vector property).
 		 */
-		DLL_IMPORT_OR_EXPORT virtual double getMinimumValue(unsigned int index = 0) const = 0;
+		DLL_IMPORT_OR_EXPORT virtual double getMinimumValue(uint64_t index = 0) const = 0;
 
 		/**
 		 * @brief	Gets the maximum value of a non vector property or the maximum value of one given
@@ -245,7 +245,7 @@ namespace RESQML2_NS
 		 * 			greater than @p 0 for a non vector property or greater than the vector size for a
 		 * 			vector property).
 		 */
-		DLL_IMPORT_OR_EXPORT virtual double getMaximumValue(unsigned int index = 0) const = 0;
+		DLL_IMPORT_OR_EXPORT virtual double getMaximumValue(uint64_t index = 0) const = 0;
 
 		/**
 		 * @brief	Sets the minimum value of a non vector property or the minimum value of one given
